+++ src/main/kotlin/pl/detailing/crm/visit/get/GetVisitDetailDtos.kt
@@ -105,7 +105,7 @@ data class ServiceLineItemResponse(
  */
 data class AdjustmentResponse(
     val type: String,
-    val value: Long
+    val value: Double  // Double to support decimal percentages like -49.19
 )
 
 /**
